adding files

+++ tutorme_backend/tutor_api/models.py
@@ -48,7 +48,7 @@ class User(AbstractUser):
 	# first_name = models.CharField(max_length=20)
 	# last_name = models.CharField(max_length=20)
 	# email = models.EmailField(max_length=254, unique=True)
-	phone = models.IntegerField()
+	phone = models.FloatField()
 	# is_admin = models.BooleanField(default=False)
 	# is_active = models.BooleanField(default=True)
 	# is_superuser = models.BooleanField(default=False)
